Changed indices to return shorts

+++ src/main/java/converter/AssignmentConverter.java
@@ -56,7 +56,7 @@ public final class AssignmentConverter implements BaseConverter {
         }
         bytes.addAll(valueConverter.convert(start));
         storeBytes.add(0, Bytecode.STORE.value);
-        storeBytes.addAll(1, Util.shortToBytes((short) info.varIndex(variable.getName())));
+        storeBytes.addAll(1, Util.shortToBytes(info.varIndex(variable.getName())));
     }
 
     private void assignToIndex(@NotNull List<Byte> bytes, List<Byte> storeBytes, int start,
